Break up results generating file to the function and hook file
