+++ api/makeResults.tsx
@@ -1,7 +1,3 @@
-import { useContext } from 'react';
-import DataContext from './DataProvider';
-import BlockchainContext from './blockchain';
-
 import 'array-flat-polyfill';
 
 interface VoteByNominee {
@@ -22,11 +18,11 @@ interface Winner {
   vote: number;
 }
 
-interface ResultsByPosition {
+export interface ResultsByPosition {
   [key: string]: Winner[];
 }
 
-function makeResults(
+export function makeResults(
   votes: VoteByNominee[],
   positions: string[],
   nominees: string[]
@@ -92,55 +88,4 @@ function makeResults(
   }, {});
 
   return [resultsByPosition, Array.from(new Set(positionsOnHold))];
-}
-
-const positions = ['pos1', 'pos2', 'pos3', 'pos4', 'pos5', 'pos6'];
-const users = ['u1', 'u2', 'u3', 'u4', 'u5', 'u6'];
-
-const votes = [
-  { vote: 8, user: 'u1', position: 'pos1' },
-  { vote: 3, user: 'u2', position: 'pos1' },
-  { vote: 2, user: 'u4', position: 'pos1' },
-  { vote: 1, user: 'u2', position: 'pos2' },
-  { vote: 5, user: 'u3', position: 'pos2' },
-  { vote: 2, user: 'u2', position: 'pos3' },
-  { vote: 4, user: 'u3', position: 'pos3' },
-  { vote: 6, user: 'u5', position: 'pos4' },
-  { vote: 6, user: 'u6', position: 'pos4' },
-  { vote: 7, user: 'u4', position: 'pos5' },
-  { vote: 7, user: 'u4', position: 'pos6' }
-];
-
-console.log(makeResults(votes, positions, users));
-
-export function useElectionResult(
-  electionId: string
-): (string[] | ResultsByPosition)[] {
-  const { blockchain } = useContext(BlockchainContext);
-  const { elections } = useContext(DataContext);
-  const { nominations } = elections.find(({ id }) => id === electionId);
-
-  const votesForElection = blockchain
-    .filter(
-      block => block.vote !== null && block.vote.electionId === electionId
-    )
-    .map(({ vote }) => vote);
-
-  const positions = Object.keys(nominations);
-  const nominees = Object.values(nominations).flat();
-
-  const votesByNominee = Object.values(nominations)
-    .map((nominees, i) => {
-      const position = positions[i];
-      return nominees.map(nominee => ({
-        user: nominee,
-        position: position,
-        vote: votesForElection.filter(
-          vote => vote.to === nominee && vote.position === position
-        ).length
-      }));
-    })
-    .flat();
-
-  return makeResults(votesByNominee, positions, nominees);
 }
